% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_data_prep.R
\name{adm_data_prep}
\alias{adm_data_prep}
\title{Admissions Data Preperation}
\usage{
adm_data_prep(dfAdm, Col, HEGIS, appVar = Applications,
  admitVar = Admitted, AcAcVar = Accepted.Offers,
  matVar = Matriculants, semYearVar = Applied.To.Year.Semester,
  colVar = Graduate.Application.College, HEGISVar = HEGIS.Code.1,
  HEGISdescVar = Graduate.Application.HEGIS)
}
\arguments{
\item{dfAdm}{a dataframe imported from ADW report A2.2}

\item{Col}{character string college abbreviation}

\item{HEGIS}{hegis number; provided as a number or character string
if there are fewer than 4 digits, a leading zero will be added}

\item{appVar}{name of column containing the count of applications. Provided either
bare (unquoted) or in open quotes. Default value is \code{Applications}}

\item{admitVar}{name of column containing the count of extedended offers. Provided either
bare (unquoted) or in open quotes. Default value is \code{Admitted}}

\item{AcAcVar}{name of column containing the count of mutual accepts (e.g. students
who accepted the offer of admission)  Provided either bare (unquoted) or in open quotes.
Default value is \code{Accepted.Offers}}

\item{matVar}{name of column containg the count of matriculated students. Provided
either bare (unquoted) or in open quotes. Default value is \code{Matriculated}}

\item{semYearVar}{name of column containing the  Semester and Year of admission.
Provided either bare (unquoted) or in open quotes. Default value is \code{Applied.To.Year.Semester}}

\item{colVar}{name of column containing the 3 character college abbreviation. Provided either
bare (unquoted) or in open quotes. Default value is \code{Graduate.Application.College}}

\item{HEGISVar}{name of column containing the HEGIS code. Provided either
bare (unquoted) or in open quotes. Default value is \code{HEGIS.Code.1}}

\item{HEGISdescVar}{name of column containing the HEGIS description. Provided either
bare (unquoted) or in open quotes. Default value is \code{Graduate.Application.HEGIS}}
}
\value{
A dataframe filtered to the college and hegis supplied with four additional columns
\code{Selectivity} is the number of admitted students divided by the total number of applicants;
\code{AppYear} is the calendar year of the semester of application;
\code{dApps} is the difference between current year and previous year apps;
\code{pApps} is the percentage change from the previous year to to the current year in applications
\code{dEnroll} is the difference between the current year and the prvious year matriculants
\code{pEnroll} is the percentage change from the previous year to the current year in matriculants
}
\description{
\code{adm_data_prep} takes an export from the ADW report A2.2
with each row representing a single year, semester, snapshot date,
and hegis; Variables are Applications, Admitted, Accepted Offers,
Declined Offers, Rejected, Registered, Matriculants.
There is no degree level or degree program;
Data should be filtered to PhD program apps prior to export
}
