% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps_offers_plotter.R
\name{apps_offers_plotter}
\alias{apps_offers_plotter}
\title{Applications & Offers Plotting}
\usage{
apps_offers_plotter(dfADM, l.appColors, n.Ynudge = NULL)
}
\arguments{
\item{dfADM}{a dataframe exported by the \code{\link{adm_data_prep}} function}

\item{l.appColors}{list of colors for the applications and offers lines; default is
c("#35618f", "#a0e3b7")}

\item{n.Ynudge}{a numeric value for adjusting the the placement of the line labels y
position, #' if not supplied, a value is calculated of 0.05 * max(# applicaitons in
latest two years)}
}
\value{
a ggplot object
}
\description{
\code{apps_offers_plotter} takes a dataframe with one application year and returns
a line plot of applicaitons and offers
}
