% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_cycle_plot_interactive.R
\name{attrition_cycle_plot_int}
\alias{attrition_cycle_plot_int}
\title{Attrition Plot Interactive}
\usage{
attrition_cycle_plot_int(dfAttrition_long, l.hegis = NULL,
  ExitPath = "TOTAL")
}
\arguments{
\item{dfAttrition_long}{a attrition dataframe formatted by the \code{\link{attrition_data_prep}} function}

\item{l.hegis}{a vector of hegis codes to plot}

\item{ExitPath}{a character string indicating which exit path ("TOTAL", "With Masters", "Without Masters"
to plot )}
}
\value{
A cycle plot attrtition data as a plotly object
}
\description{
\code{attrition_cycle_plot_int} plotting funciton for the creation of cycle plots For an overivew of cycle plots,
please see this \href{https://www.perceptualedge.com/articles/guests/intro_to_cycle_plots.pdf}{article} in
Stephen Few's Perceptual Edge library.
}
