% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_data_prep_parameterized.R
\name{attrition_data_prep_parameterized}
\alias{attrition_data_prep_parameterized}
\title{Attrition Data Preperation}
\usage{
attrition_data_prep_parameterized(dfAttrition_report, college_var = Col,
  hegiscd_var = Hegis.Cd, hegisdesc_var = Hegis.Description,
  cohort_var = Cohort.Year, WoMA_var = Without.MA..8,
  wMA_var = With.MA..9)
}
\arguments{
\item{dfAttrition_report}{a copy of the 'PhD Attrtion by Hegis Code
Template - AY[AY] thru [AY] - [date]' report imported from the xlsx}
}
\value{
A list of dataframes: one with one column with one row per program/cohort (\code{dfAttrition_wide}), and
 one with one row per program/cohort and exit pathway (\code{dfAttrition_long})

 Shared Variables:

\itemize{
\item \code{Col} is the college;
\item \code{Hegis.Cd} is HEGIS code of the phd program;
\item \code{Hegis.Description} is the HEGIS description fo the phd program;
\item \code{Cohort.Year} is the academic year of the entering cohort e.g., 02-03 is the cohort that
began in Fall 2002;
\item \code{Cohort.Size} number of students in the entering cohort;
}

\code{dfAttrition_wide} variables:
\itemize{
\item \code{Without.MA[__n]} number of students who left in a given year in the \code{n} + 1 year in program
without a Masters or PhD
\item \code{Stop.Out[__n]} number of students who left in a given year in the \code{n} + 1 year in program
but returned at a later date
\item\code{PhD.Degree[__n]} number of students who left in a given year in the \code{n} + 1 year in program
with a PhD
\item\code{Continuing[__n]} number of students who continued from year \code{n} + 1  to year \code{n} + 2
\item\code{Info.Unknown[__n]} number of students in  the \code{n} + 1 year in program that we were not
able to find information on
\item\code{Without.MA[__n]} number of students who left in a given year in the \code{n} + 1 year in program
without a Masters or PhD
\item\code{Without.MA_yr[n]} is the number of phd students who left by the end of year \code{n}
without a Masters degree or a PhD
\item\code{With.MA_yr[n]} is the number of phd students who left by the end of year \code{n}
with a Masters degree and without a PhD
 \item\code{Total_yr[n]} is the number of phd students who left by the end of year \code{n} without
 a PhD
}


\code{dfAttrition_long} variables:
\itemize{
\item\code{AcadYear} numeric variable indicating the academic year of the entering cohort (e.g., "02-03" would
have a  value of 2003)
\item\code{Pathyear} string/character variable where the first part indicates the path of exit (with MA,
without MA, total) and the second part indicates the year in program
\item\code{Count} numeric varialbe indicating the cumulative number of students who exited the program
in the given path by the year in program
\item\code{Year} string/character variable indicating the year in program
\item\code{Path} string/character variable indicating the path of exit
\item\code{Per.Cohort} string/character indicating the percentage of the cohort that exited the program
in the given path by the year in program
}
}
\description{
\code{attrition_data_prep} modifies the PhD Attrition by Hegis report
and returns a dataframe that is ready for plotting by calculatiing totals and
attrition rates
}
