% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_data_prep.R
\name{comp_data_prep}
\alias{comp_data_prep}
\title{Completions Data Preperation}
\usage{
comp_data_prep(dfCompletion_report)
}
\arguments{
\item{dfCompletion_report}{a copy of the 'PhD Completion by Hegis Code
Template - AY[AY] thru [AY] - [date]' report imported from the xlsx}
}
\value{
A data frame with one row per program/cohort and completion year. There is a
variable

\itemize{
\item \code{Col} is the college;
\item \code{Hegis.Cd} is HEGIS code of the phd program;
\item \code{Hegis.Description} is the HEGIS description fo the phd program;
\item \code{Cohort.Year} is the academic year of the entering cohort e.g., 02-03 is the cohort that
began in Fall 2002;
\item \code{Cohort.Size} number of students in the entering cohort;
\item\code{AcadYear} numeric variable indicating the academic year of the entering cohort (e.g., "02-03" would
have a  value of 2003)
\item\code{Count} numeric varialbe indicating the cumulative number of students who have completed the program
\item\code{Year} string/character variable indicating the year in program
\item\code{Per.Cohort} string/character indicating the percentage of the cohort that completed the PhD
by the Year
}
}
\description{
\code{comp_data_prep} modifies the PhD Completion by Hegis report
and returns a dataframe that is ready for plotting by calculatiing totals and
attrition rates
}
