% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{int_breaks_x}
\alias{int_breaks_x}
\title{\code{int_breaks_x} a helper function that returns a vector of integers to create
integer-only breaks for continuous x axis}
\usage{
int_breaks_x(var)
}
\arguments{
\item{var}{a continuous variable}
}
\value{
a vector of integers
}
\description{
\code{int_breaks_x} a helper function that returns a vector of integers to create
integer-only breaks for continuous x axis
}
