% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{int_breaks_y}
\alias{int_breaks_y}
\title{Helper Functions}
\usage{
int_breaks_y(var)
}
\arguments{
\item{var}{a continuous variable}
}
\value{
a vector of integers
}
\description{
\code{int_breaks_y} a helper function that returns a vector of integers to create
integer-only breaks for continuosu y axis starting at 0
}
