% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_plotter.R
\name{selectivity_plotter}
\alias{selectivity_plotter}
\title{Selectivity Plotting}
\usage{
selectivity_plotter(dfADM, l.color = c("#35618f"), theme_stripe = NULL,
  v.AppYear = AppYear, v.Selectivity = Selectivity,
  colorVar = Graduate.Application.HEGIS)
}
\arguments{
\item{dfADM}{a dataframe exported by the \code{\link{adm_data_prep}} function}

\item{l.color}{list of colors for the dots; default is "#35618f"}

\item{theme_stripe}{custom theme object for plotting}

\item{v.AppYear}{the varaible defining the year of application; default value is
\code{AppYear}}

\item{v.Selectivity}{the variable name for the selectivity metric; default value
is \code{Selectivity}}

\item{colorVar}{the variable used to define the color and fill for the dots;
the default is \code{Graduate.Applciation.HEGIS}}
}
\value{
a ggplot object
}
\description{
\code{selectivity_plotter} takes a dataframe with one application year and returns
a lollipop plot of applicaitons and offers
}
