% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttd_plotter_0-2.R
\name{ttd_plotter_2}
\alias{ttd_plotter_2}
\title{Time To Degree Plotter}
\usage{
ttd_plotter_2(dfTTD, ttdColor = "#35618f",
  ttdVar = Time.To.Degree..Yrs....22, medColor = "black")
}
\arguments{
\item{dfTTD}{a dataframe imported from}

\item{ttdColor}{fill color for the histogram or density plot; default is a deep blue: #35618f}
}
\value{
a ggplot object, if there are 10 or more rows in the supplied dataframe, then the
object returned will be a histogram. If there are 5 or more rows, but fewer than 10, the
object will be kernal desnsity plot. If there are fewer than 5 rows, the object will be
a blank plot with red text stating 'Insufficient Observations'
}
\description{
\code{ttd_plotter_2} takes a dataframe with one row per graduate
and returns a plot. If there are more than ten rows, a histogram
will be returned.
}
